requirements.txt
streamlit>=1.36.0
openai>=1.45.0
pypdf>=4.2.0
python-dotenv>=1.0.1
